+++ tests/sap/aibus/dar/client/test_data_manager_client.py
@@ -1,4 +1,5 @@
 from io import BytesIO, StringIO
+import json
 from typing import Any
 from unittest.mock import create_autospec, call, Mock, MagicMock
 
@@ -121,6 +122,51 @@ class AbstractDARClientConstruction:
         with pytest.raises(HTTPSRequired):
             # RFC 4266 URLs should also be rejected
             self.clazz.construct_from_jwt("gopher://host:70/1", jwt)
+
+    def test_create_from_cf_env(self, monkeypatch):
+        vcap_services = {
+            "data-attribute-recommendation-staging": [
+                {
+                    "binding_guid": "XXXX",
+                    "binding_name": None,
+                    "credentials": {
+                        "swagger": {
+                            "dm": self.dar_url + "data-manager/doc/ui",
+                            "inference": self.dar_url + "inference/doc/ui",
+                            "mm": self.dar_url + "model-manager/doc/ui",
+                        },
+                        "uaa": {
+                            "clientid": self.clientid,
+                            "clientsecret": self.clientsecret,
+                            "identityzone": "dar-saas-test-app",
+                            "identityzoneid": "XXX",
+                            "subaccountid": "XXX",
+                            "tenantid": "XXX",
+                            "tenantmode": "dedicated",
+                            "uaadomain": "authentication.sap.hana.ondemand.com",
+                            "url": self.uaa_url,
+                            "verificationkey": "XXX",
+                            "xsappname": "XXX",
+                            "zoneid": "XXXX",
+                        },
+                        "url": self.dar_url,
+                    },
+                    "instance_guid": "XXX",
+                    "instance_name": "dar-instance-3",
+                    "label": "data-attribute-recommendation",
+                    "name": "dar-instance-3",
+                    "plan": "standard",
+                    "provider": None,
+                    "syslog_drain_url": None,
+                    "tags": [],
+                    "volume_mounts": [],
+                }
+            ]
+        }
+
+        monkeypatch.setenv("VCAP_SERVICES", json.dumps(vcap_services))
+        client = self.clazz.construct_from_cf_env()
+        self._assert_fields_initialized(client)
 
     def _assert_fields_initialized(self, client):
         assert isinstance(client.credentials_source, OnlineCredentialsSource)